pyproject.toml
[project]
name = "web-research-agent-demo"
version = "0.3.0"
description = "Web-based research agent using Amazon Web Services Strands Agents, Ollama running Qwen3, and Tavily MCP Server."
authors = [{ name = "Gary Stafford", email = "gary.a.stafford@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.15.0",
    "python-dotenv>=1.1.1",
    "strands-agents>=1.9.1",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.8",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "types-requests>=2.30.0.0",
]

[tool.black]
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
